tests/wpt/web-platform-tests/css/css-grid/alignment/grid-self-alignment-stretch-vertical-lr-015.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Self-Alignment and stretch on auto-sized tracks</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#alignment">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-align-self">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-self">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-stretch">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-align-self-stretch">
<meta name="assert" content="The stretched orthogonal grid items along the column and/or row axis include their defined border-box.">
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<style>
.grid {
  position: relative;
  display: inline-grid;
  font: 20px/1 Ahem;
  background: grey;
  width: 250px;
  height: 250px;
  writing-mode: vertical-lr;
}
.grid > div { writing-mode: horizontal-tb; }
.grid > div { border-style: solid; }
.firstRowFirstColumn {
  grid-row: 1;
  grid-column: 1;
  background: green;
  justify-self: stretch;
  align-self: start;
  border-width: 0px 0px 10px 0px;
}
.firstRowSecondColumn {
  grid-row: 1;
  grid-column: 2;
  background: blue;
  justify-self: start;
  align-self: stretch;
  border-width: 0px 20px 0px 0px;
}
.secondRowFirstColumn {
  grid-row: 2;
  grid-column: 1;
  background: yellow;
  justify-self: start;
  align-self: start;
  border-width: 0px 0px 0px 0px;
}
.secondRowSecondColumn {
  grid-row: 2;
  grid-column: 2;
  background: red;
  justify-self: stretch;
  align-self: stretch;
  border-width: 0px 10px 20px 0px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div class="grid">
  <div data-offset-x="0"   data-offset-y="0"   data-expected-width="80"  data-expected-height="120" class="firstRowFirstColumn">XX X</div>
  <div data-offset-x="0"   data-offset-y="120" data-expected-width="130" data-expected-height="20"  class="firstRowSecondColumn">XX X</div>
  <div data-offset-x="130" data-offset-y="0"   data-expected-width="80"  data-expected-height="20"  class="secondRowFirstColumn">XX X</div>
  <div data-offset-x="130" data-offset-y="120" data-expected-width="120" data-expected-height="130" class="secondRowSecondColumn">XX X</div>
</div>
